# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Reliabilix GreenOps API
  version: 1.0.0
  description: Multi-tenant FinOps + GreenOps API. Tenancy comes from the verified identity (membership in the database) or from an API key, never
    from request input. Every tenant table is protected by Row Level Security.
servers:
- url: https://api.reliabilix.green
security:
- bearerAuth: []
...
paths:
...
  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List projects
      tags:
      - projects
      description: Requires permission `project:read`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      operationId: createProject
      summary: Create a project
      tags:
      - projects
      description: Requires permission `project:write`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{id}:
    get:
      operationId: getProject
      summary: Get a project
      tags:
      - projects
      description: Requires permission `project:read`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{id}/functional-units:
    get:
      operationId: listFunctionalUnits
      summary: Reported functional units (SCI denominator R)
      tags:
      - projects
      description: Requires permission `project:read`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FunctionalUnits'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    post:
      operationId: reportFunctionalUnits
      summary: Report functional units served in a period
      tags:
      - projects
      description: Requires permission `project:write`.
      parameters:
      - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionalUnitsInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionalUnits'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{id}/policy:
    get:
      operationId: getPolicy
      summary: Data-residency allow-list and CI thresholds
      tags:
      - projects
      description: Requires permission `project:read`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      operationId: putPolicy
      summary: Replace the policy
      tags:
      - projects
      description: Requires permission `project:write`.
      parameters:
      - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
...
  /api/v1/recommendations:
    get:
      operationId: listRecommendations
      summary: List recommendations (most carbon saved first)
      tags:
      - recommendations
      description: Requires permission `recommendation:read`.
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - open
          - approved
          - applied
          - dismissed
      - $ref: '#/components/parameters/project_id'
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/recommendations/refresh:
    post:
      operationId: refreshRecommendations
      summary: Queue a recommendation analysis for a project
      tags:
      - recommendations
      description: Requires permission `recommendation:apply`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRefInput'
      responses:
        '202':
          description: Queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queued'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/recommendations/{id}:
    get:
      operationId: getRecommendation
      summary: Get a recommendation
      tags:
      - recommendations
      description: Requires permission `recommendation:read`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/recommendations/{id}/approve:
    post:
      operationId: approveRecommendation
      summary: Approve (human decision; compliance is re-checked against the current policy)
      tags:
      - recommendations
      description: Requires permission `recommendation:apply (people only, never API keys)`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /api/v1/recommendations/{id}/apply:
    post:
      operationId: applyRecommendation
      summary: Mark as applied (only from approved)
      tags:
      - recommendations
      description: Requires permission `recommendation:apply (people only, never API keys)`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /api/v1/recommendations/{id}/dismiss:
    post:
      operationId: dismissRecommendation
      summary: Dismiss
      tags:
      - recommendations
      description: Requires permission `recommendation:apply (people only, never API keys)`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Auth0 access token (RS256) or an API key (`grk_...`). Optional header `X-Tenant-ID` selects among several memberships.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    project_id:
      name: project_id
      in: query
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
      required:
      - title
      - status
      additionalProperties: true
    Queued:
      type: object
      properties:
        status:
          type: string
          enum:
          - queued
      additionalProperties: true
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        functional_unit:
          type: string
        created_at:
          type: string
          format: date-time
      additionalProperties: true
    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        functional_unit:
          type: string
          minLength: 1
          maxLength: 60
      required:
      - name
      - functional_unit
      additionalProperties: true
    FunctionalUnits:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        units:
          type: number
      additionalProperties: true
    FunctionalUnitsInput:
      type: object
      properties:
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        units:
          type: number
          minimum: 0
          exclusiveMinimum: true
      required:
      - period_start
      - period_end
      - units
      additionalProperties: true
    Policy:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        allowed_regions:
          type: array
          items:
            type: string
          maxItems: 50
          description: 'Data-residency allow-list. Empty = residency unknown: no region_shift is proposed.'
        ci_carbon_warn_kg:
          type: number
          nullable: true
        ci_carbon_block_kg:
          type: number
          nullable: true
        ci_cost_warn:
          type: number
          nullable: true
        ci_cost_block:
          type: number
          nullable: true
      additionalProperties: true
...
    ProjectRefInput:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
      required:
      - project_id
      additionalProperties: true
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - rightsizing
          - region_shift
          - time_shift
          - spot_migration
        title:
          type: string
        provider:
          type: string
        service_name:
          type: string
        current_region:
          type: string
        recommended_region:
          type: string
        estimated_carbon_reduction_pct:
          type: number
        carbon_reduction_kg_month:
          type: number
        estimated_cost_impact:
          type: number
          nullable: true
          description: null = not estimated; negative = saving
        cost_basis:
          type: string
          enum:
          - not_estimated
          - region_price_index
        confidence:
          type: number
          minimum: 0
          maximum: 1
        compliance_check:
          type: object
          properties:
            residency:
              type: string
              enum:
              - passed
              - failed
              - unknown
            allowed_regions:
              type: array
              items:
                type: string
            checked_at:
              type: string
              format: date-time
          additionalProperties: true
        status:
          type: string
          enum:
          - open
          - approved
          - applied
          - dismissed
        decided_by:
          type: string
          nullable: true
        decided_at:
          type: string
          format: date-time
          nullable: true
        applied_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      additionalProperties: true
...